Add Markdown processing to Svelte
Created a new directory structure, pages, to separate the pages from the
routes. (You can't have extensions other than .svelte in the routes/
directory, because of a limitation in svelte-kit's route-handling code.
At least, you can't *for now*. The feature is planned for the future).

Eventually, once svelte-kit gains the ability to track other extensions
in the routes/ directory, we'll come back and merge the text of some of
the pages (such as About) back into the route, because for a simple page
like About the separation doesn't actually make sense, and it's simply
forced on us by the lack of extension configuration in svelte-kit's
routing feature at the moment. But for other pages, having the page
content in pages/ and the logic in routes/ will probably make sense, so
it's good that we can do this.