Fix load() function in blog route
There's one other difference between the load() function in Svelte Kit
and its precursor, preload(), in Sapper. In Sapper, the object returned
by preload() was an object that set the properties for the component.
But in Svelte Kit, the load() function can do more: it can return any of
the following:

{ error: { message, status }}
{ redirect }
{ context: { various properties to merge into context }}
{ maxage: (in seconds) }
{ props: { various properties for the component }}

So we just needed to return { props: { ...whatever }} and that would
have worked.